// typings.d.ts
import { Interface } from "readline";

interface CurrentWeather {
    is_day: number;
    temperature: number;
    time: string;
    weathercode: number;
    winddirection: number;
    windspeed: number;
  }

  interface Daily {
    apparent_temperature_max: number[];
    apparent_temperature_min: number[];
    sunrise: string[];
    sunset: string[];
    temperature_2m_max: number[];
    temperature_2m_min: number[];
    time: string[];
    uv_index_clear_sky_max: number[];
    uv_index_max: number[];
    weathercode: number[];
  }

interface DailyUnits {
    apparent_temperature_max:string;
    apparent_temperature_min:string;
    temperature_2m_max:string;
    temperature_2m_min:string;
    time:string;
    uv_index_clear_sky_max:string;
    uv_index_max:string;
    weathercode:string;
}

interface Hourly {
    apparent_temperature: [number];
    dewpoint_2m: [number];
    precipitation: [number];
    precipitation_probability: [number];
    rain: [number];
    relativehumidity_2m: [number];
    showers: [number];
    temperature_2m: [number];
    time: [string];
    uv_index: [number];
    uv_index_clear_sky: [number];
    visibility: [number];
    winddirection_10m: [number];
    windspeed_10m: [number];
  }

  interface HourlyUnits {
    apparent_temperature: string;
    dewpoint_2m: string;
    precipitation: string;
    precipitation_probability: string;
    rain: string;
    relativehumidity_2m: string;
    showers: string;
    temperature_2m: string;
    time: string;
    uv_index: string;
    uv_index_clear_sky: string;
    visibility: string;
    winddirection_10m: string;
    windspeed_10m: string;
  }

  interface Root {
    current_weather: CurrentWeather;
    daily: Daily;
    daily_units: DailyUnits;
    elevation: number;
    generationtime_ms: number;
    hourly: Hourly;
    hourly_units: HourlyUnits;
    latitude: number;
    longitude: number;
    timezone: string;
    timezone_abbreviation: string;
    utc_offset_seconds: number;
  }